Code/requirements.txt
accelerate==0.25.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
contourpy==1.2.0
cycler==0.12.1
datasets==2.17.1
decorator==5.1.1
deepspeed==0.12.4
dill==0.3.8
docker-pycreds==0.4.0
easydict==1.12
exceptiongroup==1.2.0
executing==2.0.1
fair-esm==2.0.0
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.42
hjson==3.1.0
huggingface-hub==0.20.3
hydra-core==1.3.2
idna==3.6
importlib_resources==6.1.2
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lit==17.0.6
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
peft==0.8.2
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
rdkit-pypi==2022.9.5
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.0.1
torch-cluster==1.6.3+pt20cu117
torch-scatter==2.1.2+pt20cu117
torch-sparse==0.6.18+pt20cu117
torch-spline-conv==1.2.2+pt20cu117
torchdrug==0.2.1
torchvision==0.15.2
tqdm==4.66.2
traitlets==5.14.1
transformers==4.35.2
triton==2.0.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.3
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0